Placed json reader within new module
